src/ingestion/requirements_fetcher.txt
google-cloud-storage==2.*
google-cloud-bigquery==3.*
google-cloud-secret-manager==2.*
requests==2.*
pandas==2.*
numpy==1.24.*